+++ objectbox/CHANGELOG.md
@@ -1,3 +1,5 @@
+## latest
+
 ## 5.0.1-dev.0 (2025-10-29)
 
 * Generator: migrate to `analyzer` 8 APIs. Require at least `analyzer` 8.1.1 and `source_gen` 4.0.1.