+++ objectbox/CHANGELOG.md
@@ -1,5 +1,10 @@
 ## latest
 
+### Sync
+
+* SyncClient: actually configure Sync filters when passing them to `Sync.client` and related helper
+  methods.
+
 ## 5.0.1-dev.0 (2025-10-29)
 
 * Generator: migrate to `analyzer` 8 APIs. Require at least `analyzer` 8.1.1 and `source_gen` 4.0.1.